Moved Investors to separete app
