+++ forum/settings.py
@@ -49,7 +49,8 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'rest_framework_simplejwt.token_blacklist',
     'users',
-    'startups'
+    'startups',
+    'investors'
 ]
 
 MIDDLEWARE = [
